requirements.txt
aiogram~=2.20
PyYAML~=6.0
aiohttp~=3.8.1
aiosqlite~=0.17.0
SQLAlchemy~=1.4.36
multidict~=6.0.2